Feature/1486 Fix small issues
* Change name of component files to upper case
* Implement local optimization modal

Resolves: #1485
