+++ src/t03/components/optimization/InputRange.jsx
