+++ src/t03/containers/modelEditorOptimization.jsx
@@ -4,19 +4,24 @@ import ConfiguredRadium from 'ConfiguredRadium';
 import {connect} from 'react-redux';
 import styleGlobals from 'styleGlobals';
 import {withRouter} from 'react-router';
-import OptimizationObjectsComponent from '../components/optimization/optimizationObjects';
-import OptimizationObjectivesComponent from '../components/optimization/optimizationObjectives';
-import OptimizationParametersComponent from '../components/optimization/optimizationParameters';
-import OptimizationConstraintsComponent from '../components/optimization/optimizationConstraints';
-import OptimizationResultsComponent from '../components/optimization/optimizationResults';
+import OptimizationObjectsComponent from '../components/optimization/OptimizationObjects';
+import OptimizationObjectivesComponent from '../components/optimization/OptimizationObjectives';
+import OptimizationParametersComponent from '../components/optimization/OptimizationParameters';
+import OptimizationConstraintsComponent from '../components/optimization/OptimizationConstraints';
+import OptimizationResultsComponent from '../components/optimization/OptimizationResults';
 import {Routing} from '../actions/index';
 import Optimization from '../../core/optimization/Optimization';
 import Stressperiods from '../../core/modflow/Stressperiods';
 import {Button, Icon, List, Menu, Popup, Progress} from 'semantic-ui-react';
 import {Action, Command} from '../actions';
 import {
-    OPTIMIZATION_STATE_CANCELLING, OPTIMIZATION_STATE_FINISHED, OPTIMIZATION_STATE_STARTED, OPTIMIZATION_STATE_CANCELLED,
-    getMessage, optimizationInProgress, optimizationHasError
+    OPTIMIZATION_STATE_CANCELLING,
+    OPTIMIZATION_STATE_FINISHED,
+    OPTIMIZATION_STATE_STARTED,
+    OPTIMIZATION_STATE_CANCELLED,
+    getMessage,
+    optimizationInProgress,
+    optimizationHasError
 } from '../selectors/optimization';
 
 const styles = {
@@ -191,6 +196,7 @@ class ModelEditorOptimization extends React.Component {
                     <OptimizationResultsComponent optimization={optimization} errors={this.state.errors}
                                                   model={this.props.model}
                                                   stressPeriods={stressPeriods}
+                                                  onChangeParameters={this.onChange}
                                                   onChange={this.onChangeResult}/>
                 );
             default:
@@ -222,12 +228,12 @@ class ModelEditorOptimization extends React.Component {
                             header='Validation Failed'
                             content={
                                 <List>
-                                    { errors.map((error, key) => (
-                                    <List.Item key={key}>
-                                        <List.Content>
-                                            {error.dataPath} {error.type} {error.message}
-                                        </List.Content>
-                                    </List.Item>
+                                    {errors.map((error, key) => (
+                                        <List.Item key={key}>
+                                            <List.Content>
+                                                {error.dataPath} {error.type} {error.message}
+                                            </List.Content>
+                                        </List.Item>
                                     ))}
                                 </List>
                             }
@@ -241,17 +247,17 @@ class ModelEditorOptimization extends React.Component {
         if (!optimizationInProgress(this.state.optimization.state)) {
             return (
                 <Menu.Item>
-                        <Button fluid primary onClick={this.onCalculationClick}>
-                            Run Optimization
-                        </Button>
+                    <Button fluid primary onClick={this.onCalculationClick}>
+                        Run Optimization
+                    </Button>
                 </Menu.Item>
             );
         }
         return (
             <Menu.Item>
-                    <Button fluid color="red" onClick={this.onCancelCalculationClick}>
-                        Cancel Calculation
-                    </Button>
+                <Button fluid color="red" onClick={this.onCancelCalculationClick}>
+                    Cancel Calculation
+                </Button>
             </Menu.Item>
         );
     }
